+++ cast/models.py
@@ -2,11 +2,11 @@ from django.db import models
 
 # Create your models here.
 class Cast(models.Model):
-	cast_id=models.PositiveSmallIntegerField()
+	cast_id=models.IntegerField()
 	cast_name=models.CharField(max_length=40,blank=False)
 	actual_name=models.CharField(max_length=40,blank=False)
 	role=models.CharField(max_length=40,blank=False)
-	image_id=models.PositiveSmallIntegerField(blank=False)
+	image_id=models.IntegerField(blank=False)
 	def __str__(self):
 		return self.cast_name
