+++ episode/models.py
@@ -2,12 +2,12 @@ from django.db import models
 
 # Create your models here.
 class Episode(models.Model):
-	episode_id=models.PositiveSmallIntegerField(blank=False)
-	episode_no=models.PositiveSmallIntegerField()
+	episode_id=models.IntegerField(blank=False)
+	episode_no=models.IntegerField()
 	episode_name=models.CharField(max_length=40)
-	server_list_id=models.PositiveSmallIntegerField()
-	views=models.PositiveIntegerField()
-	type=models.PositiveSmallIntegerField()
+	server_list_id=models.IntegerField()
+	views=models.IntegerField()
+	type=models.IntegerField()
 
 	def __str__(self):
 		return self.episode_name
