+++ genre_list/models.py
@@ -1,8 +1,8 @@
 from django.db import models
 
 class Genre_list(models.Model):
-	genre_id=models.PositiveSmallIntegerField()
-	genre_name_id=models.PositiveSmallIntegerField()
+	genre_id=models.IntegerField()
+	genre_name_id=models.IntegerField()
 
 	def __str__(self):
 		return self.genre_id
