+++ language_name/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class Language_name(models.Model):
-	language_id=models.PositiveSmallIntegerField()
+	language_id=models.IntegerField()
 	language_name=models.CharField(max_length=40)
