+++ season/models.py
@@ -2,9 +2,9 @@ from django.db import models
 
 # Create your models here.
 class Season(models.Model):
-	season_id=models.SmallIntegerField()
+	season_id=models.IntegerField()
 	season_name=models.TextField(blank=False)
-	eposide_id=models.PositiveIntegerField()
+	eposide_id=models.IntegerField()
 	plot=models.TextField()
 
 	def __str__(self):
